+++ packages/components/src/components/multi-select/multi-select-option/multi-select-option-styles.spec.ts
@@ -1,0 +1,10 @@
+import { getComponentCss } from './multi-select-option-styles';
+
+describe('getComponentCss()', () => {
+  it.each<Parameters<typeof getComponentCss>>([['light'], ['dark']])(
+    'should return correct css for theme: %s',
+    (...args) => {
+      expect(getComponentCss(...args)).toMatchSnapshot();
+    }
+  );
+});