+++ packages/components/src/components/select-wrapper/select-wrapper-dropdown/select-wrapper-dropdown-styles.ts
@@ -2,16 +2,7 @@ import type { DropdownDirectionInternal } from '../select-wrapper/select-wrapper
 import type { Theme } from '../../../types';
 import type { JssStyle, Styles } from 'jss';
 import { getCss, isHighContrastMode, mergeDeep } from '../../../utils';
-import {
-  getInsetJssStyle,
-  getTransition,
-  getThemedColors,
-  addImportantToRule,
-  hoverMediaQuery,
-  getHighContrastColors,
-  getHiddenTextJssStyle,
-  prefersColorSchemeDarkMediaQuery,
-} from '../../../styles';
+import { addImportantToRule, getInsetJssStyle, getThemedColors, getTransition, hoverMediaQuery, prefersColorSchemeDarkMediaQuery, } from '../../../styles';
 import {
   borderRadiusSmall,
   borderWidthBase,
@@ -21,7 +12,8 @@ import {
   spacingStaticSmall,
   textSmallStyle,
 } from '@porsche-design-system/utilities-v2';
-import { OPTION_HEIGHT } from '../select-wrapper/select-wrapper-styles';
+
+import { getSelectOptionStyles, OPTION_HEIGHT } from '../../../styles/select';
 import { getThemedFormStateColors } from '../../../styles/form-state-color-styles';
 import type { FormState } from '../../../utils/form/form-state';
 
@@ -216,22 +208,11 @@ export const getFilterStyles = (
 export const getListStyles = (direction: DropdownDirectionInternal, theme: Theme): Styles => {
   const isDirectionDown = direction === 'down';
   const {
-    primaryColor,
-    backgroundColor,
-    contrastMediumColor,
-    contrastHighColor,
-    backgroundSurfaceColor,
-    disabledColor,
-  } = getThemedColors(theme);
-  const {
     primaryColor: primaryColorDark,
     backgroundColor: backgroundColorDark,
     contrastMediumColor: contrastMediumColorDark,
-    contrastHighColor: contrastHighColorDark,
-    backgroundSurfaceColor: backgroundSurfaceColorDark,
-    disabledColor: disabledColorDark,
   } = getThemedColors('dark');
-  const { highlightColor } = getHighContrastColors();
+  const { primaryColor, backgroundColor, contrastMediumColor } = getThemedColors(theme);
 
   return {
     '@global': {
@@ -270,59 +251,12 @@ export const getListStyles = (direction: DropdownDirectionInternal, theme: Theme
         }),
       },
     },
-    option: {
-      display: 'flex',
-      justifyContent: 'space-between',
-      gap: '12px',
-      padding: `${spacingStaticSmall} 12px`,
-      flex: `1 0 calc(${fontLineHeight} + ${spacingStaticSmall} * 2)`,
-      color: contrastHighColor,
-      ...prefersColorSchemeDarkMediaQuery(theme, {
-        color: contrastHighColorDark,
-      }),
-      cursor: 'pointer',
-      textAlign: 'left',
-      wordBreak: 'break-word',
-      boxSizing: 'border-box',
-      borderRadius: borderRadiusSmall,
-      transition: ['background-color', 'color'].map(getTransition).join(),
-      '&[role=status]': {
-        cursor: 'not-allowed',
-      },
-      '&__sr': getHiddenTextJssStyle(),
-      ...hoverMediaQuery({
-        '&:not([aria-disabled]):not([role=status]):hover': {
-          color: isHighContrastMode ? highlightColor : primaryColor,
-          background: backgroundSurfaceColor,
-          ...prefersColorSchemeDarkMediaQuery(theme, {
-            color: isHighContrastMode ? highlightColor : primaryColorDark,
-            background: backgroundSurfaceColorDark,
-          }),
-        },
-      }),
-      '&--highlighted, &--selected': {
-        color: isHighContrastMode ? highlightColor : primaryColor,
-        background: backgroundSurfaceColor,
-        ...prefersColorSchemeDarkMediaQuery(theme, {
-          color: isHighContrastMode ? highlightColor : primaryColorDark,
-          background: backgroundSurfaceColorDark,
-        }),
-      },
+    ...getSelectOptionStyles(theme, {
       '&--selected': {
         cursor: 'default',
         pointerEvents: 'none',
       },
-      '&--disabled': {
-        cursor: 'not-allowed',
-        color: disabledColor,
-        ...prefersColorSchemeDarkMediaQuery(theme, {
-          color: disabledColorDark,
-        }),
-      },
-      '&--hidden': {
-        display: 'none',
-      },
-    },
+    }),
     optgroup: {
       display: 'block',
       padding: '3px 14px',