+++ packages/components/src/styles/form-styles.ts
@@ -116,7 +116,8 @@ export const getLabelStyles = (
   hideLabel: BreakpointCustomizable<boolean>,
   state: FormState,
   theme: Theme,
-  counterOrUnitOrIconStyles?: Styles<'counter'> | Styles<'unit'> | Styles<'icon'>
+  counterOrUnitOrIconStyles?: Styles<'counter'> | Styles<'unit'> | Styles<'icon'>,
+  additionalLabelJssStyle?: JssStyle
 ): Styles => {
   const { primaryColor, disabledColor, contrastHighColor } = getThemedColors(theme);
   const {
@@ -168,6 +169,7 @@ export const getLabelStyles = (
           },
         }),
       },
+      ...additionalLabelJssStyle,
     },
     ...(counterOrUnitOrIconStyles && {
       [counterOrUnitOrIconStylesKey]: {